+++ src/coinbase/wallet.ts
@@ -9,12 +9,20 @@ import { Address } from "./address";
 import { Coinbase } from "./coinbase";
 import { ArgumentError, InternalError } from "./errors";
 import { Transfer } from "./transfer";
-import { Amount, Destination, SeedData, WalletData, ServerSignerStatus } from "./types";
+import {
+  Amount,
+  Destination,
+  SeedData,
+  WalletData,
+  ServerSignerStatus,
+  WalletCreateOptions,
+} from "./types";
 import { convertStringToHex, delay } from "./utils";
 import { FaucetTransaction } from "./faucet_transaction";
 import { BalanceMap } from "./balance_map";
 import Decimal from "decimal.js";
 import { Balance } from "./balance";
+import { Trade } from "./trade";
 
 /**
  * A representation of a Wallet. Wallets come with a single default Address, but can expand to have a set of Addresses,
@@ -60,17 +68,23 @@ export class Wallet {
    * Instead, use User.createWallet.
    *
    * @constructs Wallet
-   * @param intervalSeconds - The interval at which to poll the CDPService, in seconds.
-   * @param timeoutSeconds - The maximum amount of time to wait for the ServerSigner to create a seed, in seconds.
+   * @param options - The options to create the Wallet.
+   * @param options.networkId - the ID of the blockchain network. Defaults to 'base-sepolia'.
+   * @param options.intervalSeconds - The interval at which to poll the CDPService, in seconds.
+   * @param options.timeoutSeconds - The maximum amount of time to wait for the ServerSigner to create a seed, in seconds.
    * @throws {ArgumentError} If the model or client is not provided.
    * @throws {InternalError} - If address derivation or caching fails.
    * @throws {APIError} - If the request fails.
    * @returns A promise that resolves with the new Wallet object.
    */
-  public static async create(intervalSeconds = 0.2, timeoutSeconds = 20): Promise<Wallet> {
+  public static async create({
+    networkId = Coinbase.networks.BaseSepolia,
+    intervalSeconds = 0.2,
+    timeoutSeconds = 20,
+  }: WalletCreateOptions = {}): Promise<Wallet> {
     const result = await Coinbase.apiClients.wallet!.createWallet({
       wallet: {
-        network_id: Coinbase.networkList.BaseSepolia,
+        network_id: networkId,
         use_server_signer: Coinbase.useServerSigner,
       },
     });
@@ -342,6 +356,23 @@ export class Wallet {
    */
   public listAddresses(): Address[] {
     return this.addresses;
+  }
+
+  /**
+   *  Trades the given amount of the given Asset for another Asset. Currently only the default address is used to source the Trade
+   *
+   * @param amount - The amount of the Asset to send.
+   * @param fromAssetId - The ID of the Asset to trade from. For Ether, eth, gwei, and wei are supported.
+   * @param toAssetId - The ID of the Asset to trade to. For Ether, eth, gwei, and wei are supported.
+   * @throws {InternalError} If the default address is not found.
+   * @throws {Error} If the private key is not loaded, or if the asset IDs are unsupported, or if there are insufficient funds.
+   * @returns The Trade object.
+   */
+  public async createTrade(amount: Amount, fromAssetId: string, toAssetId: string): Promise<Trade> {
+    if (!this.getDefaultAddress()) {
+      throw new InternalError("Default address not found");
+    }
+    return await this.getDefaultAddress()!.createTrade(amount, fromAssetId, toAssetId);
   }
 
   /**